+++ locales/en.php
@@ -1,0 +1,4 @@
+<?php
+	# ENGLISH (by default)
+	$lang['en']=array();
+?>
